Fixup the jar from the previous commit.
The jarjar project actually uses ant as its primary build tool and only
has a pom for uploading to maven central.  We don;t have authrotu for
that, so I ran `ant dist` - again on a clean r142 - and copied the
resulting jar, source jar, and javadoc jar that plopped in dist/ here.
I then renamed to match the version and generated md5 and sha1 checksum
files by hand.